requirements.txt
asgiref==3.11.1
dj-database-url==3.0.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
Pillow==11.3.0
psycopg[binary]==3.2.13
PyJWT==2.12.1
redis==6.4.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.9.0
